[FIX] account_avatax,account_external_tax: traceback with same tax in different tax lines
Set up Avatax environment
Create an Invoice with 2 lines:
- Prod A, Analytic Account A
- Prod B, Analytic Account B
Compute taxes
Now open one of the newly added taxes
On the repartition line click on the 3 dots menu, show 'Tax Closing Entry'
Disable 'Tax Closing Entry' on the invoice repartition line
Now go back to the invoice
Compute taxes again

Issue: Traceback will raise
'ValueError: Expected singleton: account.move.line(...'

This occurs because, with this tax configuration, on the move we will have 2 tax
lines for the same tax, as the lines needs a different analytic account
and cannot be grouped together.

When we compare the amount per tax we'll have a recordset and it will raise an error.

A solution is to acknowledge that we might have more than 1 line and
split the possible difference evenly on the tax lines

opw-3976597

closes odoo/enterprise#71937

X-original-commit: 3bff6e66bf80e0890134462e6560e19059005e2a
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Andrea Grazioso (agr) <agr@odoo.com>