[FIX] l10n_pe_edi: set default UoM for down payment in EDI xml
Steps to reproduce:
- Install Accounting, Sales and l10n_pe_edi
- Switch to a Peruvian company (e.g. PE Company)
- Create a SO for a Peruvian customer (e.g. Comercial Constructora)
- Confirm the SO
- Create a down payment
- Confirm the down payment
- Send it to Peru UBL 2.1

Issue:
The Electronic Service Operator (OSE) responds with an error because
an item is declared without unit of measure.

Cause:
From saas-17.1, there is no product associated to a down payment line
anymore.

Solution:
For down payment line, use the code (i.e. NIU) of the default unit of
measure (i.e. Units), which was the default UoM of the down payment
product before saas-17.1.

opw-4057252

closes odoo/enterprise#69484

X-original-commit: 5f112fe853357424447e68b416d36eb35f50f56e
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>