[FIX] timesheet_grid: fix timesheets validation domain
When the employee set on a timesheet has its timesheet_manager_id field
set to false, any user with the 'users: all timesheets' is supposed to
be able to validate/invalidate the timesheets of that employee.
Currently, this behavior is broken.

Step to reproduce:
- install timesheet on a db
- connect with a user with 'users: all timesheets' access rights
- open the timesheet app
- select the list view
- select any timesheet with an employee without an assigned manager
- validate the timesheet
An error notification is displayed to warn the user he can not validate
the timesheet.

Source of the problem:
The domain in the _get_domain_for_validation_timesheet is not taking
this use case into account.

Solution:
Update the domain in order to include the timesheet with a mangerless
employee.

task - 3922511
affected version 16.0 - master

closes odoo/enterprise#68933

Related: odoo/odoo#178342
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>