[IMP] web_studio: notification on approval/refusal
We refactor the way the field users_to_notify is used. Now, the users
to notify (if any) are notified when an action/method on a record is
approved/refused. Previously they were notified when a request was
created. They are not longer notified at that moment.

Task ID: 4122972

closes odoo/enterprise#69131

Related: odoo/upgrade#6452
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>