[IMP] web_studio,*: form view's statusbar buttons in cog menu (mobile)
*: worksheet

In a nutshell, since we now have a "CogMenu" in the ControlPanel which
also contains different actions for the current record, this change
reclaims the space taken by the "Actions" dropdown and moves those
buttons to the said CogMenu, leaving more space to the status/stage
widget.

This commit adapts tests and xpath for the changes made in community.

PS: the "isSmall" mock in worksheet was already wrong before this
commit...

task-3336242

closes odoo/enterprise#69134

Related: odoo/odoo#178091
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>