[FIX] l10n_fr_fec_import: enable importing partner with same name
Steps to reproduce:
- Have a FEC file (see ticket)
- Import it on a brand new db

Issue:
Crash. We try to create account move line with an xml id that does not exist because we didn't create the partner associated to the correct ref

The second issue is that we will link the move to the wrong customer since the second one is not created

Cause:
The first issue is caused because we bypass the creation of the second partner with the same name "PARTNER 01" but different ref "PARTNER02"
`ValueError: External ID not found in the system: l10n_fr_fec_import.89_partner_PARTNER02`

The reason is, when creating the move line, we assign the `partner_id` with an xml_id based on the partner_ref
https://github.com/odoo/enterprise/blob/ae3820059f6ff46b0dbcd64eabb2aae5caa98e18/l10n_fr_fec_import/wizard/import_wizard.py#L474-L490
Which of course does not exist.

opw-3932234

closes odoo/enterprise#69148

Signed-off-by: Julien Van Roy (juvr) <juvr@odoo.com>