[REF] spreadsheet_*: generic record name write
Previously, each implementation of `AbstractSpreadsheetAction` implemented
its own method to write on the record name.
The reason is that each model can have a different field as their name
(even if they currently are all fields named 'name').

With this commit, the abstract action handles the renaming itself.

Part-of: odoo/enterprise#69153
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>