[IMP] account_reports: improve audit of Bank in Journal Report
This commit aims to facilitate reading and handling the audit of
the Bank and Credit Card journals. This view should be tailored,
as it is not the same as suditing sales/purchase ledgers.

Beyond showing only the counterpart lines, already addressed
by #67611, the view should:
- Remove the grouping by Partner > Journal Entry
- Show columns Date, Partner and Matching

This commit adds a view for the bank/credit card audit, and
edits the function journal_report_open_aml_by_move to achieve
this tailored audit view. It also edits the List Renderer for
the audit view to ensure it can also handle the case where no
default groupings exist.

task-4147336

closes odoo/enterprise#69266

Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>