[IMP] sale_ebay: match name of variant if url mismatch
eBay recently changed the URL they send in the viewitemURL. We found 3 different ways they send it,
but there might be more. And the new ones can only be matched to the product template, not variant.
We shouldn't have relied on this URL to match products in the beginning, but we can't refactor this
flow in stable, instead, we fallback on the name of the variant, as it was created from eBay.

opw-3934127

closes odoo/enterprise#69815

X-original-commit: a0ac6660a3590b8786e75fc8f06e8369977633b6
Signed-off-by: Morgane Demesmaeker <edm@odoo.com>