[FIX] knowledge: test checking that portal user can create/edit articles
Since odoo/enterprise#38575, portal users can create/edit knowledge
articles.  However, the test that checks that feature is not correct:
 - it does not call check_access_rights() (for ACLs);
 - it uses a res.partner instead of a res.users in with_user().

Because the test only checks access rules with a non-existing user, no
rules are found, and the check is therefore trivial.  In other words,
the test passes by accident :-(

The fix consists in creating a portal user for the corresponding
"customer" contact, and add the expected calls to check_access_rights().

closes odoo/enterprise#69240

X-original-commit: f0922182c52a0337ccc57603f7f06a9f6a6329de
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>