[FIX] stock_barcode_mrp: not redirect to backend from barcode backorder
Steps to reproduce:

- Enable multi-step routes in the settings
- Inventory > Configuration > Warehouse Management > Operations types
- Click on Manufacturing and put "Always" on create a backorder
- Create a product P (no need for a bom)
- Create and confirm a manufacturing order for 2 units of the product
- Go to the barcode module > operations > manufacturing > the MO
- Register only one unit of P and produce
> A backorder is automatically created but you are redirected to the backend.

Note:

If you were in "ask" on create backorder or if your product had a bom
and you did not register the quantities of consumed components, a pop up
would appear and resolving the pop up would redirect you to the kanban
view of mrp.production in the barcode module.

Cause of the issue:

The issue was originally solved by commit d90acab by overriding the
`action_backorder` however, since commit 1e5c82f the action_backorder is
no more part of the flow in "always" flows.

opw-3890886

closes odoo/enterprise#69169

Related: odoo/odoo#178492
Signed-off-by: William Henrotin (whe) <whe@odoo.com>