[FIX] social_instagram: handle null media URL
Issue:
When clients attempt to connect their Instagram account to the Social
Marketing app, an error occurs if their account contains a post without
a media_url value (in our case, this issue involved a video).

Steps to Reproduce:
It's quite difficult to reproduce because, according to my research, it
doesn't happen with all videos. This issue might be related to the fact
that new Instagram videos have become Reels. I tried with a test Instagram
account, but I couldn't replicate this case.

Fix:
Use a default empty string instead of None (via get). This prevents the
user from encountering the error and getting blocked.

opw-4085729
opw-4141505
opw-4141479
opw-4105134
opw-4114002

closes odoo/enterprise#69373

X-original-commit: 568285d3abf71d10faa77c7ea4b6ba162fbe2f4b
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Louis Gobert (loug) <loug@odoo.com>