[FIX] *: missing entries in `assertRecordValues` expected values
Previously, `assertRecordValues` would only check the fields present
in the first "expected values" record. In `industry_fsm_stock` and
`stock_barcode_mrp` this results in a number of properties which are
set but were never tested.

Depending on test, either remove the extra values, or pin the ones
which are missing.

Part-of: odoo/enterprise#69196
Related: odoo/odoo#178332
Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>