[FIX] l10n_mx_edi{,_pos}: ensure use of company name  in CFDI
Currently, when a user requests an invoice from a POS ticket under a
Mexican company, the CFDI document attached to the invoice incorrectly
uses the buyer's name instead of using the provided company name.

Steps to reproduce
--------------------
* install `l10n_mx_edi_pos`
* Switch to a Mexican company
* the settings enable "Generate a code on ticket"
* validate a POS order
* follow the instructions on the ticket to generate an invoice. Make
  sure to provide a company name on the customer form.

When checking the XML file that is forwarded to the Mexican tax
authorities, you see that only the "individual name" is registered there
as the buyer, when it should be the company name.

opw-4103543
opw-4071503

closes odoo/enterprise#69126

Signed-off-by: Julien Van Roy (juvr) <juvr@odoo.com>