[FIX] l10n_mx_edi: missing line name description in xml
Description of the issue this commit addresses:

The xml export of invoices in the mexican edi is missing the line's description

Desired behavior after this commit is merged:

The description of the lines are present in the xml export.

Details on the fix:

The fix basically consists in using a concatenation of both the product's name
and the invoice line's name but in some configurations, the line's name might
contain the product's making it appear twice. Because of that an if-else
statement is required to check if the product's name and the invoice line's are
redundant and decides based on that.

task-4149256

closes odoo/enterprise#69219

X-original-commit: d5956f73221fac1318e7481d37f3767c26ab7560
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>