[FIX] stock_barcode, *: replace icons with text
*: mrp_workorder, pos_self_order_iot

This commit aims to add more text to the UI of the barcode app,
in case users don't know the meaning of the illustrations. A text
is shown in a button now instead.

For the IOT module, the template has been tweaked, to allow visual
consistency with the Install Kiosk action dialog. Titles are
centered horizontally.

task-4160938

closes odoo/enterprise#69855

Related: odoo/odoo#179904
Signed-off-by: Pierre Paridans (app) <app@odoo.com>