[MOV] mail: move dropzone from mail to web
In the related community PR, we moved the `Dropzone` component from the
`mail` module to the `web` module to ensure that the component can be
used in the `base_import` module. This commit will simply remove the
`mail` prefix from the classes defined in the `Dropzone` component.

COM PR: https://github.com/odoo/odoo/pull/172623

task-3956293

closes odoo/enterprise#69261

Related: odoo/odoo#172623
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>