[FIX] sale_subscription: consolidate billing by invoice address
Steps to reproduce:
- Settings > Enable 'Consolidate subscriptions billing'
- Contact > Create a company
- Contacts & Addresses tab > Add
- Add an invoice address on the company
- Add an individual then duplicate it
- Create a subscription with the indiviual as customer
- Duplicate the subscription then set the duplicate contact as customer
- Debug mode > Scheduled actions
- Sale subscription: Generate recurring invoices and payments > Run manually

The subsciptions are invoiced separately instead of having one invoice
for both. The consolidated billing should produce only one invoice since
the invoicing address is the same.

This behavior works as intended if using the create invoices action on both
subscriptions at once. Here, the recurring invoice creation for subscriptions
uses its own grouping function using partner_id as a key instead of
partner_invoice_id which is 'parent invoice address id if any, else partner_id'.

Because this grouping is performed before creating the invoices, we end
up processing sale orders that should have belonged to the same group
separarately and thus create the invoices one by one. This means we
don't let _create_invoices do the grouping since we only pass it one SO
at a time, this fix will make it so _create_invoices is called on the
grouped SOs instead.

opw-4114035

closes odoo/enterprise#69277

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>