[FIX] website_sale_subscription: allow add cart for variants recurrences
Steps
-----
[website_sale_subscription]
1. Create a recurring product with 2 recurring plans.
2. Create 2 variants for this product.
3. In the "Recurring Prices" tab, add a variant to the first plan and the
other variant to the other plan.
4. Go to eCommerce and add to the cart the second variant (not the default
selected one, this is important).
5. Refresh the page > ** This product pricing is not compatible with the
product(s) in your cart. **

Cause
-----
`_get_first_suitable_recurring_pricing` doesn't specify a plan, so it can
return a pricing which doesn't match the cart's plan.

Change
-----
Allow adding to the cart if any of the product's variants are matching the
cart's plan.

opw-4081043

closes odoo/enterprise#69288

Signed-off-by: Tanguy Quéguineur (taqu) <taqu@odoo.com>