[FIX] l10n_es_reports: do not impact all countries for generic tax
Before this commit, as soon as this module is installed, the button to
generate the Libros Registro de IVA was the main button for all the
companies.

Instead of having a new handler, just add the code in the default
handler, and manage the buttons based on the country of the current
company.

closes odoo/enterprise#69278

Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>