[FIX] spreadsheet_dashboard_account_accountant: Fix date issue
The `Accounting` dashboard was using very naive formulas to compute
deltas between years. Namely, the formulas were based on the assumpution
that we'd only compare values year by year but the filters allow to
work with quarters or months as well. This revision updates the
dashboard file to take the months/quarters into account.

opw-3957037
task-4155883

closes odoo/enterprise#69286

Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>