[IMP] web: list: select all domain when all groups are folded
This commit is an adaptation of the documents list view w.r.t. the
changes done in odoo/odoo#178779. We can now have a selection with
no visible records (the whole domain is selected in that case).
However, the delete/archive document logic requires at least a
record. We thus only enable the feature when there's one.

Task~4150362

closes odoo/enterprise#69292

Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>