[FIX] l10n_mx_edi: no line name in down payments
Description of the issue this commit addresses:

While we tought mx could not have invoices without product, it is actually
possible in down payments. Therefore a condition looking for whether there
is a product or not is missing in the invoice line name's computation.

Desired behavior after this commit is merged:

The invoice line's name always contains the name of the product just as
before but cases where there is no product are handled instead of
raising an error.

Note:

While at it, the method computing the name is renamed to better fit the helpers
present in the same file.

opw-4156804

closes odoo/enterprise#69367

X-original-commit: 7b6c19a29227ebbee080cff2351fd2a79bd5dea3
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Thomas Becquevort (thbe) <thbe@odoo.com>