[FIX] pos_blackbox_be: priority product loading
Steps to reproduce :

- Install point_of_sale
- Enable some special functionalities (ex. tip, discount etc)
- Open session with lots of products. (More than loading limit 20,000)

Issue :
Sometimes the special products required to smoothly run all functionalities
enabled are not loaded.

Cause :
We have set a limit of 20,000 and if the special product is not in those
products it won't be loaded.

Fix :
Now we ensured that all special products must be loaded without the
interference of the products  loading limit and no need to apply domain on
those 20,000 products so removed _get_available_product_domain() function.

Related PR: https://github.com/odoo/odoo/pull/175796

closes odoo/enterprise#69347

Task: 4072726
Related: odoo/odoo#175796
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>