[FIX] timesheet_grid: prevent the creation of fsm tasks without customers
Currently, when a user is creating a new task in a fsm project from the
timesheet module, a quick create is used. The issue is that unlike
standart task, fsm tasks have the field 'customer' set as required. The
quick create leave this field empty and gives no opportunity to the user
to set one even if he wanted to.

Step to reproduce:
- install industry_fsm
- open the timesheet app
- start a new timer by clicking on the 'start' button of the header
- select any fsm project
- type in the 'task' selection field to create a new one
- click on 'create'
A new fsm task without customer have benn created

Solution:
We update the 'can_create' function in order to take into account the
current project of the record. If the project is an fsm one, the
function return False. The list of fsm project is fetched on the load of
the component to avoid doing a request on every project change.

affected version 16.0 - master
task - 3922511

X-original-commit: 5fe6d00d14caa18c03ad5f9f7fba3921abbcdff1
Part-of: odoo/enterprise#69472
Related: odoo/odoo#179223
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>