[FIX] timesheet_grid: prevent creation from project user
When a user with project 'user' access rigth is creating a project on
the fly from the timesheet grid, this will trigger an access right
error. The option to create a project should not be present in the first
place for the user, since it will only lead to an access error.

step to reproduce:
- install hr_timesheet
- connect as a user with 'user' access on project module
- open the timesheet app
- start a new timesheet by clicking on the 'start' button in the header
- type a new project name inside the 'project' field
- select 'create' option
An access right error is triggered

Solution:
set the no_quick_create option to True for non admin project user.

affected version 16.0 - master
task - 3922511

X-original-commit: 22218d5e43742eb102df5f6e1f25f3950c363773
Part-of: odoo/enterprise#69472
Related: odoo/odoo#179223
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>