[IMP] account_reports: tax closing reminder activity on journal instead of move
The tax closing reminder activity used to be on the account.move for the targetted closing.

This has some drawbacks.
The main one is that we had to create an empty move when posting the current closing to be able to set the reminder on the next closing move.
But with changes made recently, we do not refresh automatically a closing move if it does already exist. This leaded to closing moves that would be empty

So this was the reason why we moved the next activity to the journal configured for closing.

task-4148136

closes odoo/enterprise#70246

Related: odoo/odoo#180629
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>