[IMP] quality_mrp_workorder: spreadsheet checks in shop floor
This commit adds the spreadsheet quality check in the shop floor.

An important current limitation is that the next check is *not* triggered
when going back from a spreadsheet check to the shop floor screen (after
click on "Save in ..."

The shop floor screen is a client action. We could export the local action
state (using the `useSetupAction` hook and `getLocalState`), restore the
state when the action is back up front when going back and trigger the
next step.

The problem is we are in a sub-sub-component when opening the spreadsheet
action. It would require a bit of refactoring to make it work and move
the state up to the client action.

However, I don't have the necessary knowledge of that code base and the
business flows to implement it quickly and well. And we are short on time.

Improving this is kept for later, probably for the mrp team which knows
better how it works.

closes odoo/enterprise#69141

Task: 4144096
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>