[FIX] appointment: Fixed the appointment slots not visible
After commit: 9e38975d3f1e9dc87a3061c06ce9df7cf562949c the appointment
page was not displaying slots, the issue was in jQuery to Vanilla JS
conversion, we were trying to get selected option using ```selected```
attribute but ```selected``` attribute was assigned on intial rendering
when we change the value it was not updated hence we have used the
```selectedIndex```.

This branch fixed the issue of vanilla JS conversion.

task-4163398

Part-of: odoo/enterprise#69389
Signed-off-by: Noé Antoine (nan) <nan@odoo.com>