[IMP] sale_subscription: make invoice_ids more resilient
When subscription is installed, there are two ways to compute the
invoices linked to a sale.order (recurring)

We can get the invoices using the relationship between account.move.line and
sale.order.line. This is the 'sale' way.

In subscription, we have a subscription_id field on the account.move.line.
Is is used in combination witht the deferred revenue functionality.

Some divergences between the computation method can be found in this
scenario:

1) create a subscription and invoice it
2) delete the sale.order.line and replace them with another product
3) the invoice_ids value is then empty because the link between the
   sale.order.line and account.move.line has been lost.

It can be mitigated by pytthon a quantity = 0 on the order line but it's
not ideal; That's why we need to take into account the subscription_id
value.

The subscription_id value could be sufficient but we prefer to keep both as they may be more resilient.

task 4163833

closes odoo/enterprise#69395

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>