[IMP] documents: no orm call on traceback setup
This commit removes the orm call which provides the id of the support
folder to the error dialog and replaces it by a simple isAdmin check.
The reason for this change is that in case of multiple tracebacks
triggered at the same time, it could overload the server with orm calls
which could also trigger more tracebacks in an infinite cycle.

closes odoo/enterprise#69406

Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>