[IMP] appointment{_account_payment}: add cancel alert to paid bookings
Before, users were sometimes lost when it came to cancelling
a paid appointment, as the validation screen did not have any
cancelling link.

In this PR, we show the link even for paid bookings. However,
we do not allow cancelling from the event summary page. Therefore,
we display an alert when trying to cancel a paid event.

The wording is improved to include name, email and phone of the
staff user. It is also updated in existing 'no-cancel' message.
That status is now 'no_time_left', a clearer name.

To do this, the controller is slightly reworked to include an
overridable method to return the cancel status.

Also, we will use in the warning message the contact info of
the event user_id (if not any inactive user). On creating the event,
the user_id is the staff_user if the appointment is based on users,
and the user who created the appointment type otherwise. As a
backup (if user is manually changed on the event for instance),
we will just fallback on 'Contact us...'.

Task-4143212

closes odoo/enterprise#69423

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>