[IMP] mrp_plm: add flexible versioning
Sometimes users don't want an ECO to affect the version number of
a product/BoM when it is applied, so we add in the option to avoid it.
By default we leave `will_update_version=True` so that the current
behavior is still the default behavior. This involved manually
updating code in several places to ensure this behavior occurs for many
different flows including:
- create BoM ECO from MO,
- applying rebase/conflict resolve due to orig BoM change or new BoM
  created by another ECO
- ECO for a product

Note that:
- the new BoM version will auto-update when the option is
  activated/deactivated so that the number matches the final applied
  version number and is easier for the user to keep track of it.
- The Update BoM (for an out of date BoM) logic has been updated to
  consider newer BoMs with the same version number (i.e. we still expect
  the previous BoMs to still be archived like the previous flow)

closes odoo/enterprise#69435

Task: 3962137
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>