[FIX] hr_appraisal_survey: Fix appraisal feedback deadline
Issue 1:
========

To reproduce:
-------------
- on an appraisal request -> ask for feedback -> set a deadline
- the employee can answer after the deadline

Problem:
--------

the method `_check_validity` was overriden in `hr_appraisal_survey` module
to do different checks than the original method in `survey` module in case
of `validity_code = answer_wrong_user`, but this override bypasses the
other checks in the original method like the deadline check.

Solution:
---------

call the original method `_check_validity` in the override, to make sure
all the checks are done, then if the original method returns
`validity_code = answer_wrong_user` we can do our specific checks.

Issue 2:
========

To reproduce:
-------------

- on an appraisal request -> ask for feedback -> set a deadline
- manager can't see the answers after the deadline

Problem:
--------

- to display answer in appraisal module we rely on the url of the survey
that contains the tokens for suervey/answer
- the deadline handling is done in the survey module where the url is dead
after the deadline which is not suitable for the appraisal module

Solution:
---------

- override the method `_get_access_data` and bypass the deadline check
in case of appraisal feedback for manager.

opw-4128638

closes odoo/enterprise#70948

X-original-commit: ffcc8c425b53df94d79ace17fb444504df8acd52
Related: odoo/odoo#181905
Signed-off-by: Abdelouahab Laaroussi (abla) <abla@odoo.com>