[FIX] web_studio: approvals: do not notify if same level rule but no entries
Have a studio approvals with minimum 3 rules:
- 2 on level 1
- 1 on level 2

Before this commit, if one of level 1 was refused, when approving the other one, the level 2 rule
created an activity for the responsible.

As we don't want a level to be notified if the lower level is not fully approved, there was a bug.

After this commit, we only trigger activity notification on higher level rules if the below level
has been fully approved.

closes odoo/enterprise#69499

Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>