[FIX] web_studio: allow approval on buttons bound to action xml_id
Have a button in a form view of type="action" and name="some.xml_id"

Before this commit, it was not possible to set approval rules on it.

This was due to code in general expecting an int, which is only part of the API
for viewButtons (see `action_service.js:doActionButton`)

After this commit, it is possible to set approval rules on such button.

backport of odoo/enterprise#69399

opw-3925360

closes odoo/enterprise#69509

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>