[FIX] knowledge: do not save virtual record on beforeUnload hook
When there is no article to load, the form view creates a virtual record
and saves it when the user closes the page or reloads it. To avoid creating
empty articles, we will trigger an urgent save only if the record is dirty.

Steps to reproduce the issue:
1. Delete all articles from the database
2. Open the Knowledge editor (you should see the no-content helper)
3. Reload the page (F5)
=> The system creates a new "Untitled" article (see the list view).
TO BE: The system should not create any article when reloading the view.

task-4167335

closes odoo/enterprise#69560

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>