[FIX] sale_subscription: remove unneeded empty <p>
Since [1] the subscriptions automations rules were extending the
view_base_automation_form view by - among other changes - displaying a
paragraph of text, but only under a certain condition.

**Before this commit**
This paragraph is present as an empty paragraph when the condition is
not met, taking up a bit of space in the form view.

**After this commit**
The empty paragraph is no more added to the DOM.

Task: 4055597

[1]: https://github.com/odoo/enterprise/commit/bbfcf24728a90b97999ce359a9d5dbf0916fa8c2

closes odoo/enterprise#69554

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>