[FIX] l10n_uy_edi: neutralize
This commit adds the missing neutralization necessary for the l10n_uy_edi
module introduced in [1]

The purpose of the standard neutralization framework is to allow us to
create database copies that will not interact with external systems in
ways that could impact the production database (or if it is not possible
to prevent the interactions, make sure that they are benign or won't
result in actual changes), or impact the customers of the operator of
the production database.

This is mainly useful to allow safe support investigation on database
duplicates.

[1] https://github.com/odoo/enterprise/pull/51266

closes odoo/enterprise#69720

X-original-commit: ab99484f8a4337aa00fc0bf5d0e7519c4b3b6998
Signed-off-by: Alexandre Moens (mao) <mao@odoo.com>
Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>