[FIX] timesheet_grid: use the TaskWithHours widget instead of the default one
This commit's purpose is to force the use of the widget 'taskWithHours'
when the timesheet timer is launched. Previously, this was done through
the xml by setting the 'type' attribute. This is no longer an option, so
instead, we update the arguments given in the getPropertyFieldInfo
method on the js side with the type we need.

task-4166816
affected version saas-17.2 - master

closes odoo/enterprise#70156

X-original-commit: edf4241e651e49627bbcf4bfb46709c67917fd43
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>