[FIX] l10n_de_reports: datev export account identifier
When exporting GL in datev format, the receivable/payable
accounts were not well formatted.

Steps:

- duplicate the Receivable Account (A)
- create a customer C with receivable set as A
- create and confirm an invoice for C
- go to GL and export datev file
-> column konto is account code of A, it should
   be 1000000 + C.id, according to datev (starting by
   1 for customers and 7 for Vendors)

Before this commit, we were making the distinction between
the customers that had the "original" receivable account and the others,
and we were displaying the account code in the second case. It should be
the datev format in any case.

opw-4117993

closes odoo/enterprise#69698

X-original-commit: f949982ee309a59370bf0ef2b61556fe12a8ce56
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Guillaume Vanleynseele (guva) <guva@odoo.com>