[FIX] web_studio: Fix export without website
This commit fixes a bug where we would always check
for some website related fields and some records which
causes an error when website was not installed.

We now simply check that website is installed before
checking the fields.

This commit does not include a test as it would require
to uninstall the website addon and the testing framework
does not permit such a thing without introducing side
effects.

closes odoo/enterprise#69575

Signed-off-by: Bruno Boi (boi) <boi@odoo.com>