[IMP] website_enterprise, website_generator: integrate website scraper
This commit integrates the Website Scraper inside the website
configurator.

This commit adds a link in the Welcome screen (1st step of the
configurator) that allows the user to go to a "Import my website"
screen. When clicking this link, the `website_generator` is installed if
it wasn't already.

This commit adds said screen with a form which, once submitted, installs
the requested modules (just `website_generator_sale` for now), creates
the generator request and goes to the WebsiteGenerator waiting screen.
Submitting the form successfully marks the configurator as done.

This commit changes the WebsiteGenerator waiting screen to use the
website loader and thus have a more unified UI.

ICP `website_generator.enable` needs to be set to true for the link to
appear on the Welcome screen and for the website generator request form
to accept submissions (`import_website` business method).

task-3339551

closes odoo/enterprise#69580

Related: odoo/odoo#179397
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>