[IMP] l10n_mx_edi_stock_30: cartaporte QR code
This commit adds a QR code (barcode) to the meixo delivery guide PDF
(cartaporte).

It is a legal requirement according to the filling guide point 4, but
other than that, it is the one thing cops will scan to check if the
document is real or not.

The QR code is generated by taking the idccp, Salida, and LLegada data
and insert them as a parameter to an URL pointing to the mexican service
to display the data of the published version of the document.

closes odoo/enterprise#71488

Task-id: 4142920
X-original-commit: 3d4d7196b03c7819a63cd709914c3f18e780d634
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Yosua Nicolaus (yoni) <yoni@odoo.com>