[FIX] partner_commission: notify progress for _cron_confirm_purchase_orders
The cron _cron_confirm_purchase_orders run only once a week, month and do nothing the rest of the week month
The cron may timeout and get delay to it's next execution date before all the po to process are done

This leave some draft pos that won't be handle before next month/week and generate unacceptable delay.

The cron should notify its progress in order to be able to run as much as needed.

In order to have a consistent progress notified, this commit first fech all the pos to process
and then process them one by one.

closes odoo/enterprise#69779

X-original-commit: b656a94084d86f9210d3cb1a6d71cd14b457399b
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Thibault Francois <tfr@odoo.com>