[FIX] l10n_ke_edi_oscu: prevent error when user upload image for create invoice
Currently, an exception is generated when the user tries to upload an
image file for creating an invoice.

error: `'utf-8' codec can't decode byte 0x89 in position 0: invalid start byte`

This is because line [1] tries to load image context of uploaded image file.

This commit fixes the above issue by handling `UnicodeDecodeError` in
`contextlib.suppress.`

[1]-https://github.com/odoo/enterprise/blob/799b80d12680153a5ba42ee97501723ba38052b1/l10n_ke_edi_oscu/models/account_move.py#L567

sentry-5757745777

closes odoo/enterprise#69649

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>