[REF] account_sepa/account_iso20022: refactor and normalize
The Canadian EFT task (3472908) was the opportunity to refactor, clean up and adjust the account_sepa module, in order to ease its usability and maintainability.

The goal is both to refactor and better the account_sepa module, which has needed a facelift for quite some time, and to ease the canadian task by making the sepa module more generic.

We first added test in the SEPA module to ensure that this refactor would be as much as possible working the same after the refactor: see https://github.com/odoo/enterprise/pull/58250

Then :
- renamed the module and the variable to be more correct regarding the nomenclature
- General clean up, de-spaghettification, simplification and removal of dead code
- Isolated the common and generic ISO20022 and created classes for each ISO20022 variants (aka pain version), avoiding as much repetition as possible 
- Modified the payment method and journals accordingly
- Replaced blocking errors and warning pop-ups by a single banner, falling back on a more generic version of ISO if the criterias aren't met
- Replaced pain.001.001.03 usage by its newer version (09), while keeping the 03 part so it can be reactivated from system settings if need be

task-3813884

closes odoo/enterprise#60393

Related: odoo/odoo#172721
Related: odoo/upgrade#5907
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>