[FIX] stock_barcode: allow old barcode compatibility
commit [1] introduced new barcode structure.
However different buisness already have printed their barcode
and it could be difficult to adapt on migration.

We have this issue. For the sake of the planet, we will add
aliases and better alias management instead
of printing in color thousands of scaleup

The demo data for scale up will be removed in 19.0

[1] commit 073163f

X-original-commit: cfe4829be00c77935555d506d86f0f14618fd21f
Part-of: odoo/enterprise#69766
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>